requirements.txt
Django==3.1
Pillow==8.1.1
djangorestframework==3.11.0
python-decouple